+++ microSWIFT_V2.2/Core/Components/Src/temp_sensor.c
@@ -151,7 +151,7 @@ static temperature_return_code_t _temperature_get_readings ( bool get_single_rea
 
 static temperature_return_code_t _temperature_start_timer ( uint16_t timeout_in_minutes )
 {
-  uint16_t timeout = TX_TIMER_TICKS_PER_SECOND * 60 * timeout_in_minutes;
+  ULONG timeout = TX_TIMER_TICKS_PER_SECOND * 60 * timeout_in_minutes;
   temperature_return_code_t ret = TEMPERATURE_SUCCESS;
 
   if ( tx_timer_change (temperature_self->timer, timeout, 0) != TX_SUCCESS )
